link PlayerController with test character controller and instantiate it, works "originize" avatar prefabs to 0,0,0

+++ WriteBetterNetcode/Assets/Scripts/Runtime/Settings/PlayerAvatarPrefabs.cs
@@ -12,6 +12,6 @@ namespace CodeSmile.Settings
 	public sealed class PlayerAvatarPrefabs : ScriptableObject
 	{
 		[SerializeField] private List<GameObject> m_Prefabs = new();
-		public GameObject this[Byte index] => index >= 0 && index < m_Prefabs.Count ? m_Prefabs[index] : null;
+		public GameObject this[int index] => index >= 0 && index < m_Prefabs.Count ? m_Prefabs[index] : null;
 	}
 }
